workers/rt_dsp/requirements.txt
dunderlab-foundation
radiant-framework
figurestream
chaski-confluent==0.1a5
numpy
dunderlab-timescaledbapp
scipy
